+++ src/SdcBundle.php
@@ -12,12 +12,14 @@
 namespace Tito10047\UX\Sdc;
 
 use Symfony\Component\Config\Definition\Configurator\DefinitionConfigurator;
+use Symfony\Component\DependencyInjection\Compiler\PassConfig;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\DependencyInjection\Extension\ExtensionInterface;
 use Symfony\Component\DependencyInjection\Extension\PrependExtensionInterface;
 use Symfony\Component\DependencyInjection\Loader\Configurator\ContainerConfigurator;
 use Symfony\Component\HttpKernel\Bundle\AbstractBundle;
 use Tito10047\UX\Sdc\CompilerPass\AssetComponentCompilerPass;
+use Tito10047\UX\Sdc\CompilerPass\ComponentAliasCompilerPass;
 use Tito10047\UX\Sdc\DependencyInjection\Configuration;
 use Tito10047\UX\Sdc\DependencyInjection\SdcExtension;
 
@@ -42,6 +44,8 @@ class SdcBundle extends AbstractBundle
     {
         parent::build($container);
 
+        // Must run before TwigComponentPass (priority 0) so alias tags are visible to it
+        $container->addCompilerPass(new ComponentAliasCompilerPass(), PassConfig::TYPE_BEFORE_OPTIMIZATION, 10);
         $container->addCompilerPass(new AssetComponentCompilerPass());
     }
 }