+++ templates/sdc/ActionTemplate.tpl.php
@@ -1,3 +1,15 @@
+<?php
+
+/*
+ * This file is part of the UX SDC Bundle
+ *
+ * (c) Jozef Môstka <https://github.com/tito10047/ux-sdc>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+?>
 {% extends 'layout.html.twig' %}
 
 {% block content %}