+++ templates/sdc/style.tpl.php
@@ -1,3 +1,5 @@
+<?php
+
 /*
  * This file is part of the UX SDC Bundle
  *
@@ -6,6 +8,8 @@
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
  */
+
+?>
 
 /* Styles for <?= $component_name ?> component */
 @layer components {